Add GPT CLI entry point and mark commands as preview
