+++ src/gpt_review/_github.py
@@ -137,7 +137,7 @@ class GitHubCommandGroup(GPTCommandGroup):
 
     @staticmethod
     def load_command_table(loader: CLICommandsLoader) -> None:
-        with CommandGroup(loader, "github", "gpt_review._github#{}") as group:
+        with CommandGroup(loader, "github", "gpt_review._github#{}", is_preview=True) as group:
             group.command("review", "_github_review", is_preview=True)
 
     @staticmethod
